requirements.txt
Django==4.0
djangorestframework==3.13.1
sendgrid-django==4.2.0